# colfetcher: use the limit hint to size the batch