@yuzefovich 2021-04-02
I think such scenario can arise when there are no stats on the table, but the query has a LIMIT.